Simplify flmexec new API usage
