+++ flmexec/Cargo.toml
@@ -6,20 +6,16 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-flame-rs = { path = "../sdk/rust" }
+flame-rs = { path = "../sdk/rust", features = ["macros"] }
 stdng = { path = "../stdng" }
 
 tokio = { workspace = true }
-tonic = { workspace = true }
 tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
 serde = { workspace = true }
-serde_json = { workspace = true }
 serde_derive = { workspace = true }
 futures = { workspace = true }
 clap = { workspace = true }
 indicatif = {version = "*", features = ["rayon"]}
-tempfile = { workspace = true }
 rand = { workspace = true }
 
 [[bin]]
